+++ etherparse/src/err/ip/headers_write_error.rs
