+++ etherparse/src/err/ip/mod.rs
@@ -1,16 +1,16 @@
-mod header_error;
-pub use header_error::*;
+mod headers_error;
+pub use headers_error::*;
 
 #[cfg(feature = "std")]
-mod header_read_error;
+mod headers_read_error;
 #[cfg(feature = "std")]
-pub use header_read_error::*;
+pub use headers_read_error::*;
 
-mod header_slice_error;
-pub use header_slice_error::*;
+mod headers_slice_error;
+pub use headers_slice_error::*;
 
-mod header_write_error;
-pub use header_write_error::*;
+mod headers_write_error;
+pub use headers_write_error::*;
 
 mod slice_error;
 pub use slice_error::*;
